Split AddressService into partial classes
