+++ RechargeSharp.v2021-11/Endpoints/Addresses/AddressService.Models.cs
@@ -1,58 +1,7 @@
-using Microsoft.Extensions.Logging;
-using RechargeSharp.v2021_11.Utilities;
-using RechargeSharp.v2021_11.Utilities.Queries;
-
 namespace RechargeSharp.v2021_11.Endpoints.Addresses;
 
-public class AddressService
+public partial class AddressService
 {
-    private readonly ILogger<AddressService> _logger;
-    private readonly IRechargeApiCaller _rechargeApiCaller;
-
-    public AddressService(ILogger<AddressService> logger, IRechargeApiCaller rechargeApiCaller)
-    {
-        _logger = logger;
-        _rechargeApiCaller = rechargeApiCaller;
-    }
-
-    public async Task<CreateAddressTypes.Response> CreateAddress(CreateAddressTypes.Request request)
-    {
-        var requestUri = $"/addresses";
-        var responseJson =
-            await _rechargeApiCaller.Post<CreateAddressTypes.Request, CreateAddressTypes.Response>(request, requestUri);
-        return responseJson;
-    }
-
-    public async Task<GetAddressTypes.Response> GetAddress(int addressId)
-    {
-        var requestUri = $"/addresses/{addressId}";
-        var responseJson = await _rechargeApiCaller.Get<GetAddressTypes.Response>(requestUri);
-        return responseJson;
-    }
-
-    public async Task<UpdateAddressTypes.Response> UpdateAddress(int addressId, UpdateAddressTypes.Request request)
-    {
-        var requestUri = $"/addresses/{addressId}";
-        var responseJson =
-            await _rechargeApiCaller.Put<UpdateAddressTypes.Request, UpdateAddressTypes.Response>(request, requestUri);
-        return responseJson;
-    }
-
-    public async Task<DeleteAddressTypes.Response> DeleteAddress(int addressId)
-    {
-        var requestUri = $"/addresses/{addressId}";
-        await _rechargeApiCaller.Delete(requestUri);
-        return new DeleteAddressTypes.Response();
-    }
-
-    public async Task<ListAddressesTypes.Response> ListAddresses(ListAddressesTypes.Request request)
-    {
-        var queryString = ObjectToQueryStringSerializer.SerializeObjectToQueryString(request);
-        var requestUri = $"/addresses{queryString.Value}";
-        var responseJson = await _rechargeApiCaller.Get<ListAddressesTypes.Response>(requestUri);
-        return responseJson;
-    }
-
     public static class SharedAddressTypes
     {
         public record ShippingLineOverride(
@@ -125,23 +74,24 @@ public class AddressService
 
     public static class UpdateAddressTypes
     {
-        public record Request(
-            string Address1,
-            string? Address2,
-            string City,
-            string? Company,
-            string CountryCode,
-            IReadOnlyList<SharedAddressTypes.Discount> Discounts,
-            string FirstName,
-            string LastName,
-            IReadOnlyList<SharedAddressTypes.OrderAttribute> OrderAttributes,
-            string? OrderNote,
-            int? PaymentMethodId,
-            string Phone,
-            string? Province,
-            IReadOnlyList<SharedAddressTypes.ShippingLineOverride> ShippingLinesOverride,
-            string Zip
-        );
+        public record Request
+        {
+            public string Address1 { get; init; }
+            public string? Address2 { get; init; }
+            public string City { get; init; }
+            public string? Company { get; init; }
+            public string CountryCode { get; init; }
+            public IReadOnlyList<SharedAddressTypes.Discount> Discounts { get; init; }
+            public string FirstName { get; init; }
+            public string LastName { get; init; }
+            public IReadOnlyList<SharedAddressTypes.OrderAttribute> OrderAttributes { get; init; }
+            public string? OrderNote { get; init; }
+            public int? PaymentMethodId { get; init; }
+            public string Phone { get; init; }
+            public string? Province { get; init; }
+            public IReadOnlyList<SharedAddressTypes.ShippingLineOverride> ShippingLinesOverride { get; init; }
+            public string Zip { get; init; }
+        }
 
         public record Response(SharedAddressTypes.Address Address);
     }
